+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## v2.0.3
 
 ### Fix bugs